# docs/openapi.yaml
openapi: 3.1.0
info:
  title: liseur-sync native API
  version: "1.0"
  description: |
    Reading-position sync and reading-statistics API.

    Conventions that apply everywhere:

    - **Auth**: native `/v1` endpoints except `/v1/login` and
      `/v1/register` require `Authorization: Bearer <token>`;
      `/healthz` is unauthenticated. OPDS uses HTTP Basic with a device
      token secret. Login returns a short-lived *auth token* usable only
      for token management; device tokens carry one or more scopes
      (`sync`, `read-insights`, `library-read`, `library-manage`,
      `library-upload`, `library-delete`, `admin`).
    - **Folder access**: catalog scopes and folder grants are independent.
      Catalog listings contain only books from folders explicitly granted to
      the token's account. Direct access to an inaccessible folder, book or
      entity returns `404`; administrator status does not bypass this rule.
    - **Identity**: books are *works*. Clients never need a work to
      exist before pushing: call `/v1/works/resolve` with every
      identifier you have and use the returned `work_id`.
    - **Idempotency**: `POST /v1/ops` and `POST /v1/sessions` are safe
      to retry. Replaying the same `op_id`/`session_id` with the same
      payload reports `duplicate`; the same id with a different payload
      is a conflict (never an overwrite).
    - **Delta sync**: `GET /v1/changes?since=<seq>` streams every op
...
      4xx with a reason, never a 5xx.
...
servers:
...
paths:
...
  /v1/login:
    post:
      tags: [auth]
      security: []
      summary: Exchange credentials for a short-lived auth token
      description: |
        The auth token (1 h) authorises token management only; it cannot
        sync or read insights. Requires HTTPS unless the instance sets
        `insecure_http`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  auth_token: { type: string, description: "Short-lived (1 h) token-management credential" }
                  expires_in: { type: integer, description: seconds }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error", description: "HTTPS required" }
        "429": { $ref: "#/components/responses/Error", description: "Rate limited" }
  /v1/register:
    post:
      tags: [auth]
      security: []
      summary: Create an account with an invite code
      description: Registration is invite-only. Invite codes are single-use and expire.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite, username, password]
              properties:
                invite: { type: string }
                username: { type: string }
                password: { type: string, minLength: 8 }
                timezone:
                  type: string
                  description: IANA name. Omitted or invalid values use UTC.
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: { type: string }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error", description: "Invalid or expired invite" }
        "409": { $ref: "#/components/responses/Error", description: "Username taken" }
  /v1/token:
    get:
      tags: [auth]
      summary: Describe the calling token
      description: |
        Returns the device token that made the request: its id, its
        device id, its name and its scopes. Call it once at startup and
        draw the interface from the answer, rather than probing routes
        and reading `403`s.

        Authenticated like every other route — an absent, revoked or
        expired token gets `401` — but it requires no particular scope.
        The subject of the response is the credential presenting it, so
        the narrowest token can ask about itself.

        It never returns a secret, a secret hash, or any token other
        than the caller's. To enumerate the account's tokens, use
        `GET /v1/tokens`, which needs the login credential.

        `HEAD` is served as the status line of the `GET`, which is what
        makes it a cheap liveness check for a stored credential. Every
        other method is `405`.

        Like every authenticated route, a call updates the token's
        last-used timestamp. Nothing about the credential's authority —
        its scopes, device binding, expiry or secret — changes.
      security: [{ deviceToken: [] }]
      responses:
        "200":
          description: The calling token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenSelf" }
        "401": { $ref: "#/components/responses/Error" }
    head:
      tags: [auth]
      summary: Check the calling token is still live
      description: |
        The status line of the `GET`, with no body. The cheapest way for
        a client holding a stored secret to find out whether it is still
        good before using it: `200` if the token authenticates, `401` if
        it is absent, revoked or expired.
      security: [{ deviceToken: [] }]
      responses:
        "200": { description: The token is live }
        "401": { $ref: "#/components/responses/Error" }
  /v1/tokens:
    post:
      tags: [auth]
      summary: Create a device token
      description: |
        Requires the login auth token (not a device token). The token
        secret is returned once; store it. Each token is bound to a
        server-generated `device_id` that stamps every op and session it
        pushes.

        `admin` scope may only be requested by a caller who already
        holds an active admin token; otherwise the request is refused
        with `403`. The first admin token on an instance is minted with
        `liseur-sync admin mint-token -scope admin`.

        Legacy clients may send `scope`. New clients send `scopes`.
        Supplying both is accepted only when they describe the same
        singleton set. Responses always return `scopes` and return the
        deprecated `scope` field only for singleton sets.
      security: [{ loginAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              anyOf:
                - required: [scope]
                - required: [scopes]
              properties:
                name: { type: string, example: "Boox Palma" }
                scope:
                  allOf:
                    - { $ref: "#/components/schemas/Scope" }
                  deprecated: true
                  description: "Legacy singleton form."
                scopes: { $ref: "#/components/schemas/ScopeSet" }
                expires_in_seconds: { type: integer }
      responses:
        "201":
          description: Token created; secret shown once
          content:
            application/json:
              schema:
                type: object
                required: [token_id, device_id, name, scopes, secret, expires_at]
                properties:
                  token_id: { type: string }
                  device_id: { type: string }
                  name: { type: string }
                  scope:
                    allOf:
                      - { $ref: "#/components/schemas/Scope" }
                    deprecated: true
                    description: "Present only for singleton scope sets."
                  scopes: { $ref: "#/components/schemas/ScopeSet" }
                  secret: { type: string, description: "Bearer secret, returned exactly once" }
                  expires_at: { type: [string, "null"], format: date-time }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403":
          description: |
            `admin` scope requested by a caller who is not already an
            admin.
          content:
            application/json:
              schema:
                type: object
                properties: { error: { type: string } }
    get:
      tags: [auth]
      summary: List your tokens
      security: [{ loginAuth: [] }]
      responses:
        "200":
          description: Token list (secrets never included)
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokens:
                    type: array
                    items: { $ref: "#/components/schemas/TokenInfo" }
  /v1/tokens/{id}:
    patch:
      tags: [auth]
      summary: Update a device token's scopes in place
      description: |
        Preserves the token id, device id, and secret. Requires the
        short-lived login credential. Adding `admin` requires the user
        to already hold an active admin token.
      security: [{ loginAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              anyOf:
                - required: [scope]
                - required: [scopes]
              properties:
                scope:
                  allOf:
                    - { $ref: "#/components/schemas/Scope" }
                  deprecated: true
                scopes: { $ref: "#/components/schemas/ScopeSet" }
      responses:
        "200":
          description: Token scopes updated in place
          content:
            application/json:
              schema:
                type: object
                required: [status, scopes]
                properties:
                  status: { type: string, const: updated }
                  scope:
                    allOf:
                      - { $ref: "#/components/schemas/Scope" }
                    deprecated: true
                    description: "Present only for singleton scope sets."
                  scopes: { $ref: "#/components/schemas/ScopeSet" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [auth]
      summary: Revoke a token
      security: [{ loginAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/StatusOK" }
        "404": { $ref: "#/components/responses/Error" }
  /v1/works/resolve:
    post:
      tags: [identity]
      summary: Resolve a file's identifiers to a work
      description: |
        Send every identifier you have for the file. Resolution priority is
        `sha256` → `partial-md5` → `source` → `dc` → `ta`, but every supplied
        identifier that already matches must agree on one work. On a
        high-confidence result, all supplied identifiers are registered as
        aliases and every supplied SHA-256 is registered as an edition of the
        resolved work (this is how identity converges across re-encodes).
        Fuzzy `ta` matches return `confidence: low` — confirm with the
        user before treating it as authoritative. On a low match nothing
        is registered, so a guess cannot be laundered into a certainty;
        send `confirmed: true` once the user has said it is the same
        book, and the match resolves high and registers the identifiers.

        If the identifiers resolve to more than one distinct work, the
        server answers 409 with the conflicting work ids and changes
        nothing; call `/v1/works/merge` to confirm the merge.
      security: [{ deviceToken: [] }]  # requires sync scope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifiers]
              properties:
                identifiers:
                  type: array
                  minItems: 1
                  items: { $ref: "#/components/schemas/Identifier" }
                title: { type: string }
                author: { type: string }
                confirmed:
                  type: boolean
                  description: Confirm that a low-confidence title/author match is the same work.
      responses:
        "200":
          description: Resolved to an existing work
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResolveResult" }
        "201":
          description: New work created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResolveResult" }
        "400": { $ref: "#/components/responses/Error" }
        "409":
          description: Identifiers map to multiple works; nothing changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  works: { type: array, items: { type: string } }
  /v1/works/{id}/split:
    post:
      tags: [identity]
      summary: Detach an edition into a new work
      description: |
        Repairs a wrong merge. Moves the edition and its ops/sessions to
        a new work. Aliases belong to works, not editions, so pass the
        aliases to move explicitly (the edition's own `sha256` alias is
        implied). A different edition's `sha256` alias is rejected.
        Works with compacted session rollups cannot be split because the
        aggregate no longer retains enough edition provenance. The same
        restriction applies to legacy inferred sessions whose edition or
        origin alias cannot be recovered.
      security: [{ deviceToken: [] }]  # requires sync scope
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [edition_sha]
              properties:
                edition_sha: { type: string }
                aliases:
                  type: array
                  items: { $ref: "#/components/schemas/Identifier" }
      responses:
        "200":
          description: Split complete
          content:
            application/json:
              schema:
                type: object
                properties: { work_id: { type: string } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error", description: "Edition or history cannot be split from this work" }
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      description: Scoped device token (`sync`, `read-insights`, `library-read`, `library-manage`, `library-upload`, `library-delete`, or `admin`)
    loginAuth:
      type: http
      scheme: bearer
      description: Short-lived auth token from /v1/login (token management only)
    opdsBasic:
      type: http
      scheme: basic
      description: |
        OPDS only. Username is the token's name or the literal `token`;
        the password is a device token secret carrying `library-read`.
        Account passwords are rejected — a reader stores its credential
        in plain text on the device.
...